+++ internal/app/model/testing.go
@@ -2,10 +2,9 @@ package model
 
 import "testing"
 
-// TestStudent ...
 func TestStudent(y *testing.T) *Student {
 	return &Student{
-		TelegramID: 99999,
+		TelegramID: int64(99999),
 		FirstName:  "FirstName",
 		LastName:   "LastName",
 		Username:   "Username",
